+++ common/src/main/java/dev/aika/smsn/config/SMSNConfig.java
@@ -9,7 +9,7 @@ import java.nio.file.Files;
 import java.util.*;
 
 @Getter
-@SuppressWarnings({"unused", "SpellCheckingInspection"})
+@SuppressWarnings("unused")
 public class SMSNConfig extends ModConfig {
     //NOTE: Not adding the @LoaderSpecific annotation means the field applies to all Mod Loaders.
     //NOTE: Not adding the @Category annotation means using the "general" category.
@@ -55,6 +55,8 @@ public class SMSNConfig extends ModConfig {
     public boolean unionLibSupporterCheck = false;
     @LoaderSpecific(LoaderType.NEOFORGE)
     public boolean alexModsContributorCheck = false;
+    @LoaderSpecific(LoaderType.NEOFORGE)
+    public boolean voidscapeDonator = false;
 
     @Category("qol")
     @Components.Switch(checked = "enable", unchecked = "disabled")